mistral/agents/agents_api/travel_assistant/pyproject.toml
[project]
name = "travel-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "google-search-results>=2.4.2",
    "ipykernel>=6.29.5",
    "loguru",
    "mistralai>=1.9.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
]


[tool.uv.workspace]
members = ["travel_assisstant"]
